Extended database interface
- moved sample data generation to test database
- added support to get Ingredients of Recipe
- added test for those
