+++ src/main/java/sample/model/IRecipe.java
@@ -2,6 +2,8 @@ package sample.model;
 
 import sample.model.activejdbc.Recipe;
 
+import java.util.Map;
+
 /**
  * Created by czoeller on 01.04.16.
  */
@@ -12,8 +14,10 @@ public interface IRecipe {
 
     boolean saveIt();
     String getTitle();
+    long getID();
     void setTitle(String title);
     void add(IRecipeIngredient recipeIngredient);
-
     void add(String ingredientName, int amount, String unitName);
+    Map<IIngredient, Map<Integer, IUnit>> getIngredients();
+    byte[] getImage();
 }
