+++ src/main/java/sample/model/activejdbc/Recipe.java
@@ -1,5 +1,6 @@
 package sample.model.activejdbc;
 
+import org.javalite.activejdbc.LazyList;
 import org.javalite.activejdbc.Model;
 import org.javalite.activejdbc.annotations.BelongsTo;
 import org.javalite.activejdbc.annotations.BelongsToParents;
@@ -7,7 +8,9 @@ import org.javalite.activejdbc.annotations.Many2Many;
 import org.javalite.activejdbc.annotations.Table;
 import sample.model.*;
 
+import java.util.Map;
 import java.util.Optional;
+import java.util.TreeMap;
 
 /**
  * Created by czoeller on 25.03.16.
@@ -43,17 +46,40 @@ public class Recipe extends Model implements IRecipe {
 
         // Create Many2Many Relation Recipe<---recipeIngredient--->Ingredient
         final RecipeIngredient recipeIngredient = RecipeIngredient.createIt("amount", amount);
-
-        // set recipeIngredient ends
-        // Create ingredient on the fly if it was not there yet
-        ingredient.orElseGet(() -> (IIngredient) Ingredient.createIt("name", ingredientName + "else")).add(recipeIngredient);
-        //ingredient.orElseGet( Ingredient.createIt("name", ingredientName + "else") ).add(recipeIngredient);
-        this.add((Model) recipeIngredient);
-
         // Create HasMany Relation unit ---< recipe_ingredient
         // Create unit on the fly if it was not there yet
         unit.orElseGet(() -> (IUnit) Unit.createIt("name", unitName)).add(recipeIngredient);
+
+        // set both recipeIngredient ends
+        // Create ingredient on the fly if it was not there yet
+        ingredient.orElseGet(() -> (IIngredient) Ingredient.createIt("name", ingredientName)).add(recipeIngredient);
+        this.add((Model) recipeIngredient);
+
     }
+
+    @Override
+    public Map<IIngredient, Map<Integer, IUnit>> getIngredients() {
+        Map<IIngredient, Map<Integer, IUnit>> map = new TreeMap<>();
+        Map<Integer, IUnit> innerMap = new TreeMap<>();
+
+        final LazyList<RecipeIngredient> recipeIngredients = this.getAll(RecipeIngredient.class);
+
+        for( RecipeIngredient recipeIngredient : recipeIngredients ) {
+            final IIngredient ingredient = recipeIngredient.parent(Ingredient.class);
+            final Integer amount = recipeIngredient.getInteger("amount");
+            final IUnit unit = recipeIngredient.parent(Unit.class);
+            innerMap.put(amount, unit);
+            map.put(ingredient, innerMap);
+        }
+
+        return map;
+    }
+
+    @Override
+    public byte[] getImage() {
+        return getBytes("image");
+    }
+
 
     public String getTitle() {
         return getString("title");
@@ -62,5 +88,11 @@ public class Recipe extends Model implements IRecipe {
     public void setTitle(String title) {
         set("title", title);
     }
+
+    @Override
+    public long getID() {
+        return this.getLongId();
+    }
+
 
 }
